[ch11999] - Backoff library may cause panic on 32bit linux devices
